// src/generated/api/models/notificationIndex200DataUnreadItem.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Módulo predefinido
 * OpenAPI spec version: 1.0.0
 */

export type NotificationIndex200DataUnreadItem = {
  task_id?: string;
  title?: string;
  due_summary?: string;
  is_read?: boolean;
  /** @nullable */
  read_at?: string | null;
  current_due_date?: string;
  days_until_due?: number;
};